+++ src/main/java/DAO/daoInterface/RouteDAO.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface RouteDAO {
     void addRoute(Route route);
     List<Route> findAllRouts();
-    Route findRouteByID(int id);
-    void setRouteRelevant(int id);
+    Route findRouteByID(Long id);
+    void setRouteRelevant(Long id);
     List<Route> findAllBetweenTwoStations(String startStation, String arrivalStation, LocalDate data, LocalTime localTime);
     List<Route> findAllBetweenTwoCites(String startCity, String arrivalCity, LocalDate data, LocalTime localTime);
